rename: 토큰 API 파일명 변경
