+++ src/app/(team)/team/[teamid]/tasklist/_components/ModalButton.tsx
@@ -1,0 +1,15 @@
+'use client';
+
+export default function ModalButton() {
+  const handleModalOpen = () => {};
+
+  return (
+    <button
+      type="button"
+      onClick={handleModalOpen}
+      className="bg-[#000000] text-[#ffffffff]"
+    >
+      모달
+    </button>
+  );
+}
